+++ packages/klingai/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/klingai
 
+## 4.0.0-canary.39
+
+### Patch Changes
+
+- Updated dependencies [d848405]
+  - @ai-sdk/provider-utils@5.0.0-canary.37
+
 ## 4.0.0-canary.38
 
 ### Patch Changes