+++ packages/langchain/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @ai-sdk/langchain
 
+## 3.0.0-canary.130
+
+### Patch Changes
+
+- ai@7.0.0-canary.130
+
+## 3.0.0-canary.129
+
+### Patch Changes
+
+- Updated dependencies [d1b3786]
+  - ai@7.0.0-canary.129
+
+## 3.0.0-canary.128
+
+### Patch Changes
+
+- ai@7.0.0-canary.128
+
 ## 3.0.0-canary.127
 
 ### Patch Changes