+++ packages/open-responses/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/open-responses
 
+## 2.0.0-canary.40
+
+### Patch Changes
+
+- Updated dependencies [d848405]
+  - @ai-sdk/provider-utils@5.0.0-canary.37
+
 ## 2.0.0-canary.39
 
 ### Patch Changes