+++ packages/rsc/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/rsc
 
+## 3.0.0-canary.131
+
+### Patch Changes
+
+- ai@7.0.0-canary.130
+
+## 3.0.0-canary.130
+
+### Patch Changes
+
+- Updated dependencies [d1b3786]
+  - ai@7.0.0-canary.129
+
+## 3.0.0-canary.129
+
+### Patch Changes
+
+- Updated dependencies [d848405]
+  - @ai-sdk/provider-utils@5.0.0-canary.37
+  - ai@7.0.0-canary.128
+
 ## 3.0.0-canary.128
 
 ### Patch Changes