+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/svelte
 
+## 5.0.0-canary.130
+
+### Patch Changes
+
+- ai@7.0.0-canary.130
+
+## 5.0.0-canary.129
+
+### Patch Changes
+
+- Updated dependencies [d1b3786]
+  - ai@7.0.0-canary.129
+
+## 5.0.0-canary.128
+
+### Patch Changes
+
+- Updated dependencies [d848405]
+  - @ai-sdk/provider-utils@5.0.0-canary.37
+  - ai@7.0.0-canary.128
+
 ## 5.0.0-canary.127
 
 ### Patch Changes